fixed down script on swc foreign key

+++ migrations/950107_settlementWindowContent_foreign_settlementModel.js
@@ -23,28 +23,38 @@
  --------------
  ******/
 
-'use strict'
+ 'use strict'
 
-exports.up = async (knex) => {
-  return await knex.schema.hasTable('settlementWindowContent').then(async function (exists) {
-    if (exists) {
-      await knex.schema.hasColumn('settlementWindowContent', 'settlementModelId')
-        .then(async (columnExists) => {
-          if (!columnExists) {
-            return knex.schema.alterTable('settlementWindowContent', (t) => {
-              t.integer('SettlementModelId').unsigned().defaultTo(null)
-              t.foreign('SettlementModelId').references('SettlementModelId').inTable('settlementModel')
-            })
-          }
-        })
-    }
-  })
-}
-
-exports.down = function (knex) {
-  return knex.schema
-    .alterTable('settlementWindowContent', (t) => {
-      t.dropForeign('SettlementModelId')
-      t.dropColumn('SettlementModelId')
-    })
-}
+ exports.up = async (knex) => {
+   return await knex.schema.hasTable('settlementWindowContent').then(async function (exists) {
+     if (exists) {
+       await knex.schema.hasColumn('settlementWindowContent', 'settlementModelId')
+         .then(async (columnExists) => {
+           if (!columnExists) {
+             return knex.schema.alterTable('settlementWindowContent', (t) => {
+               t.integer('SettlementModelId').unsigned().defaultTo(null)
+               t.foreign('SettlementModelId').references('SettlementModelId').inTable('settlementModel')
+             })
+           }
+         })
+     }
+   })
+ }
+ 
+ exports.down = function (knex) {
+   return await knew.schema.hasTable('settlementWindowContent').then(async function (exists) {
+     if (exists) {
+       await knex.schema.hasColumn('settlementWindowContent', 'settlementModelId')
+         .then(async (columnExists) => {
+           if (columnExists) {
+             return knex.schema
+               .alterTable('settlementWindowContent', (t) => {
+                 t.dropForeign('SettlementModelId')
+                 t.dropColumn('SettlementModelId')
+               })
+           }
+         })
+     }
+   })
+ }
+
